Refactored DB interface
