+++ backend/src/impl/db_utils/db_utils.py
@@ -1,4 +1,5 @@
 from collections.abc import Callable
+from dataclasses import dataclass
 from typing import Optional, TypeVar
 
 from bson.objectid import InvalidId, ObjectId
@@ -9,92 +10,111 @@ from pymongo.cursor import Cursor
 from pymongo.results import DeleteResult, InsertManyResult, InsertOneResult
 
 
-class DBModel:
+@dataclass
+class DBCollection:
+    db_name: str
+    collection_name: str
+
+
+class DBUtils:
     _database_name: str
     _collection_name: str
 
-    @classmethod
-    def get_database(cls):
-        return get_db().cx[cls._database_name]
+    # Names of DBs or collections
+    SYSTEM_OUTPUT_DB = "system_outputs"
+    DATASET_METADATA = DBCollection(
+        db_name="metadata", collection_name="dataset_metadata"
+    )
+    DEV_SYSTEM_METADATA = DBCollection(
+        db_name="metadata", collection_name="dev_system_metadata"
+    )
 
-    @classmethod
-    def get_client(cls) -> ClientSession:
+    @staticmethod
+    def get_database(db_name: str):
+        return get_db().cx[db_name]
+
+    @staticmethod
+    def get_client() -> ClientSession:
         return get_db().db.client
 
-    @classmethod
-    def get_collection(cls, collection_name: str, check_collection_exist=True):
-        database = cls.get_database()
+    @staticmethod
+    def get_collection(collection: DBCollection, check_collection_exist=True):
+        database = DBUtils.get_database(collection.db_name)
         if check_collection_exist:
             collection_names = database.list_collection_names()
-            if collection_name in collection_names:
-                return database.get_collection(collection_name)
-            raise DBModelException(f"collection: {collection_name} does not exist")
-        return database.get_collection(collection_name)
+            if collection.collection_name in collection_names:
+                return database.get_collection(collection.collection_name)
+            raise DBUtilsException(
+                f"collection: {collection.collection_name} does not exist"
+            )
+        return database.get_collection(collection.collection_name)
 
-    @classmethod
-    def drop(cls, check_collection_exist=False):
+    @staticmethod
+    def drop(collection: DBCollection, check_collection_exist=False):
         """
         drop a collection
         :param check_collection_exist: if True and collection doesn't exist, raise
               exception
         """
-        if not cls._collection_name:
-            raise DBModelException("collection_name not defined")
-        cls.get_collection(cls._collection_name, check_collection_exist).drop()
+        if not collection.collection_name:
+            raise DBUtilsException("collection_name not defined")
+        DBUtils.get_collection(collection, check_collection_exist).drop()
 
-    @classmethod
+    @staticmethod
     def insert_one(
-        cls,
+        collection: DBCollection,
         document: dict,
         check_collection_exist: bool = True,
         session: ClientSession = None,
     ) -> InsertOneResult:
-        if not cls._collection_name:
-            raise DBModelException("collection_name not defined")
-        return cls.get_collection(
-            cls._collection_name, check_collection_exist
-        ).insert_one(document, session=session)
+        if not collection.collection_name:
+            raise DBUtilsException("collection_name not defined")
+        return DBUtils.get_collection(collection, check_collection_exist).insert_one(
+            document, session=session
+        )
 
-    @classmethod
+    @staticmethod
     def insert_many(
-        cls,
+        collection: DBCollection,
         documents: list[dict],
         check_collection_exist=True,
         session: ClientSession = None,
     ) -> InsertManyResult:
-        if not cls._collection_name:
-            raise DBModelException("collection_name not defined")
-        return cls.get_collection(
-            cls._collection_name, check_collection_exist
-        ).insert_many(documents, session=session)
+        if not collection.collection_name:
+            raise DBUtilsException("collection_name not defined")
+        return DBUtils.get_collection(collection, check_collection_exist).insert_many(
+            documents, session=session
+        )
 
-    @classmethod
-    def find_one_by_id(cls, docid: str, projection: Optional[dict] = None):
+    @staticmethod
+    def find_one_by_id(
+        collection: DBCollection, docid: str, projection: Optional[dict] = None
+    ):
         """
         Find and return a document with the _id field
         Prameters:
           - id: value of _id
           - projection: include or exclude fields in the document
         """
-        if not cls._collection_name:
-            raise DBModelException("collection_name not defined")
         try:
-            return cls.get_collection(cls._collection_name).find_one(
+            return DBUtils.get_collection(collection).find_one(
                 {"_id": ObjectId(docid)}, projection
             )
         except InvalidId:
             abort_with_error_message(400, f"id: {docid} is not a valid ID")
 
-    @classmethod
-    def delete_one_by_id(cls, docid: str, session: ClientSession = None):
+    @staticmethod
+    def delete_one_by_id(
+        collection: DBCollection, docid: str, session: ClientSession = None
+    ):
         """
         Delete one document with the given ID
         Returns: `True` if a single document has been deleted
         """
-        if not cls._collection_name:
-            raise DBModelException("collection_name not defined")
+        if not collection.collection_name:
+            raise DBUtilsException("collection_name not defined")
         try:
-            result: DeleteResult = cls.get_collection(cls._collection_name).delete_one(
+            result: DeleteResult = DBUtils.get_collection(collection).delete_one(
                 {"_id": ObjectId(docid)}, session=session
             )
             if int(result.deleted_count) == 1:
@@ -104,17 +124,17 @@ class DBModel:
         except InvalidId:
             abort_with_error_message(400, f"id: {docid} is not a valid mongodb ID")
 
-    @classmethod
-    def count(cls, filt: dict = None) -> int:
+    @staticmethod
+    def count(collection: DBCollection, filt: dict = None) -> int:
         if filt is None:
             filt = {}
-        if not cls._collection_name:
-            raise DBModelException("collection_name not defined")
-        return cls.get_collection(cls._collection_name).count_documents(filt)
+        if not collection.collection_name:
+            raise DBUtilsException("collection_name not defined")
+        return DBUtils.get_collection(collection).count_documents(filt)
 
-    @classmethod
+    @staticmethod
     def find(
-        cls,
+        collection: DBCollection,
         filt: Optional[dict] = None,
         sort: Optional[list] = None,
         skip=0,
@@ -137,45 +157,41 @@ class DBModel:
           - a number that represents the total matching documents without considering
             skip/limit
         """
-        if not cls._collection_name:
-            raise DBModelException("collection_name not defined")
+        if not collection.collection_name:
+            raise DBUtilsException("collection_name not defined")
 
         if not filt:
             filt = {}
-        cursor = cls.get_collection(cls._collection_name).find(filt, projection)
+        cursor = DBUtils.get_collection(collection).find(filt, projection)
         if sort:
             cursor = cursor.sort(sort)
 
         cursor = cursor.skip(skip).limit(limit)
-        total = cls.count(filt)
+        total = DBUtils.count(collection, filt)
         return cursor, total
 
     CallbackRetType = TypeVar("CallbackRetType")
 
-    @classmethod
+    @staticmethod
     def execute_transaction(
-        cls, callback: Callable[[ClientSession], CallbackRetType]
+        callback: Callable[[ClientSession], CallbackRetType]
     ) -> CallbackRetType:
         """
         Executes `callback` in a transaction. Returns the return of callback. An
         exception is raised if failure.
         - Ref: https://pymongo.readthedocs.io/en/stable/api/pymongo/client_session.html
         """
-        with cls.get_client().start_session() as session:
+        with DBUtils.get_client().start_session() as session:
             with session.start_transaction():
                 return callback(session)
 
 
-class MetadataDBModel(DBModel):
-    _database_name = "metadata"
-
-
-class DBModelException(Exception):
+class DBUtilsException(Exception):
     """
     An exception type originated from `db_model`. This exception usually means that a
     developer is not using `db_model` properly. The details of the exception is usually
     not useful for the users.
-      - `DBModelException` should not be used if a user requests for a system (with id)
+      - `DBUtilsException` should not be used if a user requests for a system (with id)
       that does not exist. For this purpose, use `abort_with_error()` or define some
       other exception type.
     """
