+++ backend/src/impl/default_controllers_impl.py
@@ -12,20 +12,26 @@ from explainaboard import (
     get_task_categories,
 )
 from explainaboard import metric as exb_metric
-from explainaboard.feature import BucketInfo, Dict, Position, Sequence, Value
+from explainaboard.feature import Dict, Position, Sequence, Value
 from explainaboard.info import SysOutputInfo
 from explainaboard.loaders.loader_registry import get_supported_file_types_for_loader
 from explainaboard.metric import MetricStats
 from explainaboard.processors.processor_registry import get_metric_list_for_processor
 from explainaboard.utils.tokenizer import get_default_tokenizer
+from explainaboard.utils.typing_utils import unwrap
 from explainaboard_web.impl.auth import get_user
-from explainaboard_web.impl.db_models.dataset_metadata_model import DatasetMetaDataModel
-from explainaboard_web.impl.db_models.system_metadata_model import (
-    SystemModel,
-    SystemOutputsModel,
-)
+
+# from explainaboard_web.impl.db_models.dataset_metadata_model import DatasetMetadata
+# from explainaboard_web.impl.db_models.system_metadata_model import (
+#     SystemModel,
+#     SystemOutputsModel,
+# )
+from explainaboard_web.impl.db_utils.dataset_db_utils import DatasetDBUtils
+from explainaboard_web.impl.db_utils.db_utils import DBUtils
+from explainaboard_web.impl.db_utils.system_db_utils import SystemDBUtils
 from explainaboard_web.impl.private_dataset import is_private_dataset
 from explainaboard_web.impl.utils import abort_with_error_message, decode_base64
+from explainaboard_web.models import DatasetMetadata
 from explainaboard_web.models.datasets_return import DatasetsReturn
 from explainaboard_web.models.system import System
 from explainaboard_web.models.system_analyses_return import SystemAnalysesReturn
@@ -84,10 +90,11 @@ def tasks_get() -> list[TaskCategory]:
 """ /datasets """
 
 
-def datasets_dataset_id_get(dataset_id: str) -> DatasetMetaDataModel:
-    dataset = DatasetMetaDataModel.find_one_by_id(dataset_id)
-    if not dataset:
+def datasets_dataset_id_get(dataset_id: str) -> DatasetMetadata:
+    document = DBUtils.find_one_by_id(DBUtils.DATASET_METADATA, dataset_id)
+    if not document:
         abort_with_error_message(404, f"dataset id: {dataset_id} not found")
+    dataset = DatasetMetadata.from_dict(document)
     return dataset
 
 
@@ -99,7 +106,7 @@ def datasets_get(
     page_size: int,
 ) -> DatasetsReturn:
     parsed_dataset_ids = dataset_ids.split(",") if dataset_ids else None
-    return DatasetMetaDataModel.find(
+    return DatasetDBUtils.find_datasets(
         page, page_size, parsed_dataset_ids, dataset_name, task
     )
 
@@ -107,8 +114,8 @@ def datasets_get(
 """ /systems """
 
 
-def systems_system_id_get(system_id: str) -> SystemModel:
-    system = SystemModel.find_one_by_id(system_id)
+def systems_system_id_get(system_id: str) -> System:
+    system = DBUtils.find_one_by_id(DBUtils.DEV_SYSTEM_METADATA, system_id)
     if not system:
         abort_with_error_message(404, f"system id: {system_id} not found")
     return system
@@ -138,7 +145,7 @@ def systems_get(
 
     dir = ASCENDING if sort_direction == "asc" else DESCENDING
 
-    return SystemModel.find(
+    return SystemDBUtils.find_systems(
         ids,
         page,
         page_size,
@@ -152,7 +159,7 @@ def systems_get(
     )
 
 
-def systems_post(body: SystemsBody) -> SystemModel:
+def systems_post(body: SystemsBody) -> System:
     """
     aborts with error if fails
     TODO: error handling
@@ -173,7 +180,7 @@ def systems_post(body: SystemsBody) -> SystemModel:
         body.system_output.data = decode_base64(body.system_output.data)
         if body.custom_dataset and body.custom_dataset.data:
             body.custom_dataset.data = decode_base64(body.custom_dataset.data)
-        system = SystemModel.create(
+        system = SystemDBUtils.create_system(
             body.metadata, body.system_output, body.custom_dataset
         )
         return system
@@ -189,7 +196,7 @@ def systems_system_id_outputs_get(
     """
     TODO: return special error/warning if some ids cannot be found
     """
-    sys = SystemModel.find_one_by_id(system_id)
+    sys = DBUtils.find_one_by_id(DBUtils.DEV_SYSTEM_METADATA, system_id)
     if not sys:
         abort_with_error_message(400, "system does not exist")
     if is_private_dataset(
@@ -202,11 +209,12 @@ def systems_system_id_outputs_get(
         abort_with_error_message(
             403, f"{sys.system_info.dataset_name} is a private dataset", 40301
         )
-    return SystemOutputsModel(system_id).find(output_ids, limit=10)
+
+    return SystemDBUtils.find_system_outputs(system_id, output_ids, limit=10)
 
 
 def systems_system_id_delete(system_id: str):
-    success = SystemModel.delete_one_by_id(system_id)
+    success = SystemDBUtils.delete_system_by_id(system_id)
     if success:
         return "Success"
     abort_with_error_message(400, f"cannot find system_id: {system_id}")
@@ -228,7 +236,7 @@ def systems_analyses_post(body: SystemsAnalysesBody):
     page = 0
     page_size = len(system_ids)
     sort = None
-    systems: list[System] = SystemModel.find(
+    systems: list[System] = SystemDBUtils.find_systems(
         system_ids,
         page,
         page_size,
@@ -280,10 +288,7 @@ def systems_analyses_post(body: SystemsAnalysesBody):
                 else feature.bucket_info
             )
 
-            if bucket_info is not None:
-                bucket_info = BucketInfo(**bucket_info)
-
-            feature.bucket_info = bucket_info
+            feature.bucket_info = unwrap(bucket_info)
             # user-defined bucket info
             if feature_name in custom_feature_to_bucket_info:
                 custom_bucket_info = custom_feature_to_bucket_info[feature_name]
@@ -316,11 +321,12 @@ def systems_analyses_post(body: SystemsAnalysesBody):
 
         # Get the entire system outputs
         output_ids = None
-        system_outputs = (
-            SystemOutputsModel(system.system_id)
-            .find(output_ids, limit=0)
-            .system_outputs
-        )
+        system_outputs = [
+            x.to_dict()
+            for x in SystemDBUtils.find_system_outputs(
+                system.system_id, output_ids, limit=0
+            ).system_outputs
+        ]
 
         fine_grained_statistics = processor.get_fine_grained_statistics(
             system_output_info,
